Fix code scanning alert #4: Incorrect conversion between integer types

+++ server/model/system/Search.go
@@ -718,7 +718,7 @@ func GetSearchPage(s SearchVo) []SearchInfo {
 	if s.Language != "" {
 		query = query.Where("language = ?", s.Language)
 	}
-	if int(s.Year) > time.Now().Year()-12 {
+	if s.Year > int64(time.Now().Year()-12) {
 		query = query.Where("year = ?", s.Year)
 	}
 	switch s.Remarks {
